Sygnały prawie zrobione, tylko SIGCONT/TSTP bash?

+++ pm.c
@@ -9,6 +9,7 @@
 #include <signal.h>
 #include <sys/shm.h>
 #include <sys/ipc.h>
+#include <semaphore.h>
 #define BSIZE 256
 
 void main_handler(int signal, siginfo_t* info,void*);
@@ -18,6 +19,13 @@ pid_t child[3];
 
 int main()
 {
+    //ustawienie semaforow p1 i p3
+    int semMemID=shmget(IPC_PRIVATE,2*sizeof(sem_t),IPC_CREAT|0666);
+    sem_t *sem = (sem_t*)shmat(semMemID,NULL,0);
+    mutex_rd=sem;
+    mutex_wr=sem+1;
+    sem_init(mutex_rd,1,0);
+    sem_init(mutex_wr,1,0);
     //ustawienie maski sygnalow
     sigfillset(&blocked);
     sigdelset(&blocked,SIGTERM);
@@ -38,23 +46,29 @@ int main()
         mkfifo(fifoLoc[i],0666);
     }
     //ustawianie pipe
-    int pipeMemID;
-    if((pipeMemID=shmget(IPC_PRIVATE,3*2*sizeof(int),IPC_CREAT|0666))==-1)
+    int pCountID = shmget(IPC_PRIVATE,sizeof(int),IPC_CREAT|0666);
+    if(pCountID==-1)
     {
-        printf("Shm error\n");
+        printf("Error shmget\n");
+        return 4;
     }
-    pipefd[0]=(int*)shmat(pipeMemID,NULL,0);
-    pipefd[1]=pipefd[0]+2;
-    pipefd[2]=pipefd[1]+2;
-    for(int i=0;i<3;i++)
+    pipe_counter = (int*)shmat(pCountID,NULL,0);
+    if(pipe_counter==NULL)
     {
-        if(pipe(pipefd[i])==-1)
+        printf("Error shmat\n");
+        return 5;
+    }
+    *pipe_counter=0;
+    for(int a=0;a<1024;a++)
+    {
+        for(int i=0;i<3;i++)
         {
-            printf("Error pipe %d\n",i);
+            if(pipe(pipefd[a][i])==-1)
+            {
+                printf("Error pipe [%d][%d]\n",a,i);
+            }
         }
     }
-
-
 
     //tworzenie procesow potomnych
     
@@ -93,14 +107,15 @@ int main()
     //macierzysty
     struct sigaction sa_pm;
     sigfillset(&sa_pm.sa_mask);
-    sa_pm.sa_flags=SA_SIGINFO;
+    sa_pm.sa_flags=SA_SIGINFO|SA_RESTART;
     sa_pm.sa_sigaction = main_handler;
     sigaction(SIGTSTP,&sa_pm,NULL);
     sigaction(SIGTERM,&sa_pm,NULL);
     sigaction(SIGCONT,&sa_pm,NULL);
+    sigaction(SIGUSR1,&sa_pm,NULL);
     for(int i=0;i<3;i++) wait(&process_ret[i]);
     int ret_val[3];
-    
+    shmdt(pipe_counter);
 
     for(int i=0;i<3;i++)
     {
@@ -116,20 +131,31 @@ int main()
 void main_handler(int signal,siginfo_t *info, void* ptr)
 {
     pid_t senderPID = info->si_pid;
+    const int sigRecv = signal;
     if(senderPID==child[2])
     {
-        printf("SenderPID=%d, signal=%d\n",senderPID,signal);
-        for(int i=0; i<3; i++)
+        if(sigRecv!=SIGUSR1)
         {
-            close(pipefd[i][0]);
-            if(write(pipefd[i][1],&signal,sizeof(int))==-1)
+            //printf("SenderPID=%d, signal=%d\n",senderPID,signal);
+            for(int i=0; i<3; i++)
             {
-                printf("Pipe write error\n");
-            }
-            close(pipefd[i][1]);
-        } 
-        printf("Wpisano do pipe\n");
-        kill(child[0],SIGUSR1);
-        printf("SIGUSR1 wyslany\n");
+                close(pipefd[*pipe_counter][i][0]);
+                if(write(pipefd[*pipe_counter][i][1],&sigRecv,sizeof(int))==-1)
+                {
+                    printf("Pipe write error\n");
+                }
+                close(pipefd[*pipe_counter][i][1]);
+            } 
+            //printf("Wpisano do pipe\n");
+            kill(child[0],SIGUSR1);
+            //printf("SIGUSR1 wyslany\n");
+        }
+        else
+        {
+            printf("Wstrzymano\n");
+            //execl("/bin/bash","");
+            sigsuspend(&blocked);
+        }
+        
     }
 }
